Split the Databricks test into unit test and integration test
